fix patch

+++ Jenkinsfile
@@ -76,7 +76,7 @@ pipeline {
                     if [ ! -f ~/.ssh/id_ed25519 ]; then
                         ssh-keygen -t ed25519 -f ~/.ssh/id_ed25519 -N ""
                     fi
-                    rm ~/.ssh/known_hosts
+                    [ -f ~/.ssh/known_hosts ] && rm ~/.ssh/known_hosts || true
                 '''
             }
         }
